+++ src/main/java/com/uet/auction/common/DTO/AuctionDTO.java
@@ -1,19 +1,20 @@
 package com.uet.auction.common.DTO;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class AuctionDTO implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private int id;
     private String productName;
-    private double currentPrice;   // Giá cao nhất hiện tại
-    private double myBid;          // Giá mà User này đã đặt
+    private BigDecimal currentPrice;   // Giá cao nhất hiện tại
+    private BigDecimal myBid;          // Giá mà User này đã đặt
     private String status;         // "Dẫn đầu", "Bị vượt mặt", "Kết thúc"
     private String timeLeft;       // Thời gian còn lại (định dạng HH:mm:ss)
     private String imagePath;      // Đường dẫn ảnh sản phẩm
 
-    public AuctionDTO(int id, String productName, double currentPrice, double myBid, String status, String timeLeft, String imagePath) {
+    public AuctionDTO(int id, String productName, BigDecimal currentPrice, BigDecimal myBid, String status, String timeLeft, String imagePath) {
         this.id = id;
         this.productName = productName;
         this.currentPrice = currentPrice;
@@ -26,8 +27,8 @@ public class AuctionDTO implements Serializable {
     // Getter và Setter cho tất cả các trường
     public int getId() { return id; }
     public String getProductName() { return productName; }
-    public double getCurrentPrice() { return currentPrice; }
-    public double getMyBid() { return myBid; }
+    public BigDecimal getCurrentPrice() { return currentPrice; }
+    public BigDecimal getMyBid() { return myBid; }
     public String getStatus() { return status; }
     public String getTimeLeft() { return timeLeft; }
     public String getImagePath() { return imagePath; }
